MocoTrack with full muscles
In this step I have run the MocoTrack on the simplified model (reduced DOF ) with all important knee flexors and extensor muscles (Millard2012). It's worthwhile to mention I couldn't run this with DeGroot muscle model because this model produces unreasonable force for the passive part when hip flexion is 90 and knee is 0. To describe simulation: isokinetic low speed knee flexion without muscle's activation (a=0.01). there is actuator on knee which is responsible to provide torque in the way that knee flexes with constant speed.